Move pool module to internal
