+++ internal/pool/metric_pool.go
