+++ pkg/statsd/lexer.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 
 	"github.com/atlassian/gostatsd"
-	"github.com/atlassian/gostatsd/pkg/pool"
+	"github.com/atlassian/gostatsd/internal/pool"
 )
 
 type lexer struct {
